NOT A WORKING COMMIT, fixing issues raised in code review
